refactor: application 레이어 삭제
